+++ reportportal_client/_internal/services/statistics.py
@@ -22,9 +22,9 @@ import aiohttp
 import certifi
 import requests
 
-from reportportal_client.helpers import get_package_parameters
 from reportportal_client._internal.services.client_id import get_client_id
 from reportportal_client._internal.services.constants import CLIENT_INFO, ENDPOINT
+from reportportal_client.helpers import get_package_parameters
 
 logger = logging.getLogger(__name__)